Add Makefile and .gitignore.
The Makefile is quite sweet and makes installing Vim packages quite easy.
The .gitignore file is used to exclude the file config.mk, as this file is
only really to be used on a per-user basis – it contains site/user-specific
configurations of the default values for where to install and what programs
to use and so on.